move index route into mixin
